fixed incorrect definition of locale in typescript
the first parameter of locale is optional so that locale() can be called with no arguments to return the current locale.
I would have added a test case to prove this works, but everything is in javascript and I expect you would not appreciate adding typescript to the unit tests.